pyproject.toml
[project]
name = "bbook-maker"
description = "The bilingual_book_maker is an AI translation tool that uses ChatGPT to assist users in creating multi-language versions of epub/txt files and books."
readme = "README.md"
license = {text = "MIT"}
dynamic = ["version"]
requires-python = ">=3.10"
authors = [
    { name = "yihong0618", email = "zouzou0208@gmail.com" },
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "anthropic",
    "backoff",
    "bs4",
    "ebooklib>=0.19",
    "google-generativeai",
    "langdetect",
    "litellm",
    "openai>=1.68.2",
    "PyDeepLX",
    "requests",
    "rich>=14.0.0",
    "tiktoken",
    "tqdm",
    "groq>=0.5.0",
    "promptdown>=0.9.0",
    "fastapi>=0.115.14",
    "beautifulsoup4>=4.13.4",
    "pytest>=8.4.1",
    "lxml>=6.0.0",
    "html5lib>=1.1",
    "soupsieve>=2.7",
    "python-dotenv>=1.1.1",
]

[project.scripts]
bbook_maker = "book_maker.cli:main"
dir_process = "utils.process_md_dir:main"
rename_files = "utils.rename_files:main"
promptdown = "promptdown_cli:main"

[project.urls]
Homepage = "https://github.com/yihong0618/bilingual_book_maker"

[project.optional-dependencies]
fastapi = [
    "uvicorn>=0.35.0",
    "python-multipart>=0.0.20",
]
[tool.pdm]
plugins = ["pdm-autoexport"]
[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = true
[build-system]
requires = ["pdm-backend>=2.0.0"]
build-backend = "pdm.backend"
[tool.pdm.version]
source = "scm"
